I aligned the cactus border along the bottom for both mobile and desktop.
Jay, I also got rid of the red border at the bottom. It wasn't very difficult. I will have to work more on figuring out how to make it meet with the cactus border. For now, it does not meet. I might also make another image for the bottom border.